Weird stuff when I tried merging but BIG ADDITIONS!
Added Writer+Reader classes in a new package! These will be used in
autonomous to record the xbox controller input to a text file in a
readable format to be read later! I thought it was much better than
having to interpret encoder data and do all sorts of math to actually
find out what to tell the robot to do so I came up with this geniusness
which is 110% accurate and realiable. I'm almost done with the Writer
class and will start the Reader class soon! !WARNING! DO NOT UNDER ANY
CIRCUMSTANCE CHANGE THIS CODE! IT TOOK ME ALL NIGHT TO FINE TUNE THIS
AND I DO PLAN ON USING IT LATER FOR ANOTHER PROJECT OF MINE! or i will
hunt you down and make you pay :angry: